Refactor release card description
