+++ release-prep/lib/release_card_description/base.rb
@@ -1,0 +1,19 @@
+module ReleaseCardDescription
+  class Base
+    def self.build(release:)
+      new(release:).build
+    end
+
+    def initialize(release:)
+      @release = release
+    end
+
+    def build
+      raise NotImplementedError, "Subclass must implement build"
+    end
+
+    private
+
+    attr_reader :release
+  end
+end
