+++ release-prep/lib/release_card_description/feature_flags_table.rb
@@ -1,0 +1,16 @@
+require_relative "base"
+
+module ReleaseCardDescription
+  class FeatureFlagsTable < Base
+    def build
+      <<~MARKDOWN
+        h2. Referenced Environment Feature Flags
+
+        ||Environment Feature Flage||Enabled||
+        #{release.environment_feature_flags.map do |feature|
+          "|#{feature}|false|"
+        end.join("\n")}
+      MARKDOWN
+    end
+  end
+end
