+++ release-prep/lib/release_card_description/header.rb
@@ -1,0 +1,15 @@
+require_relative "base"
+
+module ReleaseCardDescription
+  class Header < Base
+    def build
+      "h1. #{summary}"
+    end
+
+    private
+
+    def summary
+      "#{ENV.fetch('GITHUB_REPO')} #{release.version.name}"
+    end
+  end
+end
